[refactor]: 회원가입 로직 훅으로 추출 (#58)
